requirements.txt
Django>=5.2.7
django-cors-headers
django-filter
djangorestframework
